💥 use enum values and toml table for config units
**Breaking Change**

I was hesitant to implement this. For the reason that the application
won't run if a user has modified his config file and modified values
to not match enum variants.
This is also the case if the execution is based on CLI arguments solely,
and does not depend on values in the config file.

For such cases, I thought it'd be better to execute anyway and use the
broken options default values.
However, this also makes it harder for the user to spot misconfiguration
and guide him to the desired result.

In the end, I think it's a good and also commonly used practice that
an application won't start and throw an error if the config contains
unallowed values.

On top, I'm about to integrate more configuration options which I want
to approach it in a similar fashion using structs and enum variants.
So here we go.