♻️ forecast as option (#38)
improves accessibility, but reduces the further extensibility of the
forecast subcommand. Considering the rather small end product this app
aims to be, this is a fair trade.

this change can be seen as a breaking change as the weather forecast
is now accessed without flags using:
`-f w` | `-f d` | `-f w,d`
`-f` solely or e.g. `-f -w` is not possible anymore